refactor(core): split 4 files + rename 2 modules for NLOC compliance and domain naming
NLOC splits (all under 500 now):
- conditions.rs (568→361): extract conditions_specialized.rs (213) for similarity/geo/contains parsers
- config.rs (556→378): integrate config_quantization.rs (192) for QuantizationConfig domain
- native_index.rs (549→452): extract native_index_io.rs (107) for persistence I/O
- log_payload.rs (541→451): extract log_payload_io.rs (104) for WAL record format

Domain naming improvements:
- pipeline_helpers.rs → query_pipeline.rs (orchestration, not "helpers")
- crud_bulk_raw.rs → bulk_import.rs (import domain, not "raw")
