+++ crates/velesdb-core/src/collection/core/bulk_import.rs
